@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	"github.com/fatih/color"
+	"github.com/urfave/cli/v2"
 	"github.com/zalando/go-keyring"
 
 	"github.com/tonedefdev/terracreds/api"
@@ -18,9 +19,19 @@ import (
 type Linux struct{}
 
 // Create stores or updates a Terraform API token in Gnome Keyring or an external vault provider
-func (l *Linux) Create(cfg api.Config, hostname string, token interface{}, user *user.User, vault vault.TerraVault) error {
+func (l *Linux) Create(cfg api.Config, hostname string, token any, user *user.User, vault vault.TerraVault) error {
 	var method string
 	method = "Updated"
+
+	if token == nil {
+		var response api.CredentialResponse
+		err := json.NewDecoder(os.Stdin).Decode(&response)
+		if err != nil {
+			helpers.CheckError(err)
+		}
+
+		token = response.Token
+	}
 
 	if vault != nil {
 		_, err := vault.Get()
@@ -42,15 +53,6 @@ func (l *Linux) Create(cfg api.Config, hostname string, token interface{}, user 
 	_, err := keyring.Get(hostname, string(user.Username))
 	if err != nil {
 		method = "Created"
-	}
-
-	if token == nil {
-		err = json.NewDecoder(os.Stdin).Decode(&api.CredentialResponse{})
-		if err != nil {
-			helpers.CheckError(err)
-		}
-		err = keyring.Set(hostname, string(user.Username), api.CredentialResponse{}.Token)
-		return err
 	}
 
 	str := fmt.Sprintf("%v", token)
@@ -161,4 +163,36 @@ func (l *Linux) Get(cfg api.Config, hostname string, user *user.User, vault vaul
 
 	fmt.Fprintf(color.Output, "%s: You do not have permission to view this credential\n", color.RedString("ERROR"))
 	return nil, err
+}
+
+func (l *Linux) List(c *cli.Context, cfg api.Config, secretNames []string, user *user.User, vault vault.TerraVault) ([]string, error) {
+	var secretValues []string
+	if cfg.Logging.Enabled == true {
+		msg := fmt.Sprintf("- user requesting access: %s", string(user.Username))
+		helpers.Logging(cfg, msg, "INFO")
+	}
+
+	if vault != nil {
+		secrets, err := vault.List(secretNames)
+		if err != nil {
+			return nil, err
+		}
+
+		return secrets, nil
+	}
+
+	for _, secret := range secretNames {
+		if cfg.Logging.Enabled == true {
+			msg := fmt.Sprintf("- secret name requested: %s", secret)
+			helpers.Logging(cfg, msg, "INFO")
+		}
+
+		cred, err := keyring.Get(secret, string(user.Username))
+		if err == nil {
+			value := string(cred)
+			secretValues = append(secretValues, value)
+		}
+	}
+
+	return secretValues, nil
 }
